pyproject.toml
[tool.poetry]
name = "fast_elm"
version = "0.1.0"
description = "A fast data logger for ELM327."
authors = ["Pedro von Hertwig Batista <pedrovhb@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10,<3.11"
loguru = "^0.6.0"
rich = "^12.5.1"
black = {extras = ["d"], version = "^22.8.0"}
pyserial = "^3.5"
zstandard = "^0.18.0"
pyserial-asyncio = "^0.6"
typer = "^0.6.1"
requests = "^2.28.1"
setuptools-scm = "^7.0.5"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.19.0"
pytest = "^7.1.3"
bump2version = "^1.0.1"
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
ordered-set = "^4.1.0"


#[tool.mypy]
#mypy_path = "./"
#python_version = "3.10"
#strict = true
#
#warn_unreachable = true
#
#follow_imports = "silent"
#warn_redundant_casts = true
#warn_unused_ignores = true
#disallow_any_generics = true
#check_untyped_defs = true
#implicit_reexport = true
#disallow_untyped_defs = false

# plugins = ["pydantic.mypy"]
# [tool.pydantic-mypy]
# init_forbid_extra = true
# init_typed = true
# warn_required_dynamic_aliases = true
# warn_untyped_fields = true
mypy = "^0.971"

[tool.poetry.scripts]
fast_elm = "fast_elm.main:app"
fe = "fast_elm.main:app"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
